Feat: Updated argument validation to include extra information.
Now it is possible to configure more information using ArgsValidation annotation. It is possible to configure each argument name, description, position, validator and if it is an optional argument or not. Besides improving the overall functionality of the library, this also allows plugins using DevCommand to easily obtain information about the required arguments for the registered commands. With this information, for example, plugins can easily create Help command pages.